requirements.txt
fastapi
uvicorn
openai

langchain
langchain-community
langchain-openai
langchain-text-splitters

chromadb
python-dotenv
tiktoken